backend/app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
httpx
python-dotenv
redis>=5.0.0
fastapi-limiter
aioredis
pytest
pytest-asyncio
httpx
